# docs/platform-api-v2.openapi.yaml
openapi: 3.1.0
info:
  title: AD Platform Participant API
  version: 2.0.0
  summary: Participant-facing API for attack-defense automation
  description: >
    Participant API for target discovery, whitebox source download, flag
    submission, service unlock, SSH access, and service recovery workflows.
servers:
  - url: https://api.example.com
    description: Replace with the real participant API host
...
paths:
...
  /api/v2/game/status:
    get:
      tags:
        - Discovery
      summary: Get public game status
      description: Get public match state, scheduler state, and current tick.
      operationId: getGameStatus
      responses:
        '200':
          description: Public game status snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatus'
              examples:
                running:
                  value:
                    match:
                      state: running
                      started_at: 2026-03-10T10:00:00Z
                      accepting_submissions: true
                    current_tick:
                      id: 248
                      status: completed
                    scheduler:
                      state: running
                      interval_seconds: 60
                      last_tick_id: 248
                    total_ticks: 248
                finished:
                  value:
                    match:
                      state: finished
                      started_at: 2026-03-10T10:00:00Z
                      ended_at: 2026-03-10T12:00:00Z
                      accepting_submissions: false
                    current_tick:
                      id: 250
                      status: completed
                    scheduler:
                      state: stopped
                      interval_seconds: 60
                      last_tick_id: 250
                    total_ticks: 250
        '429':
          description: Game status polling rate limited
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /api/v2/submit:
    post:
      tags:
        - Submission
      summary: Submit flags
      description: >
        Submit one or more candidate flags. The request succeeds at the HTTP
        layer when the batch is processed; each flag receives its own result.
      operationId: submitFlags
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitFlagsRequest'
            examples:
              default:
                value:
                  flags:
                    - FLAGv1.foo.bar
                    - FLAGv1.baz.qux
      responses:
        '200':
          description: Submission batch processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResult'
              examples:
                default:
                  value:
                    results:
                      - flag: FLAGv1.foo.bar
                        status: invalid
                        detail: flag is wrong or expired.
                      - flag: FLAGv1.baz.qux
                        status: accepted
                        detail: flag is correct.
                      - flag: FLAGv1.baz.qux
                        status: duplicate
                        detail: flag already submitted.
                    accepted_count: 1
                    rejected_count: 2
        '400':
          description: Submission rejected before processing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalidPayload:
                  value:
                    title: Invalid request
                    status: 400
                    detail: request body is invalid.
                notStarted:
                  value:
                    title: Submission rejected
                    status: 400
                    detail: contest has not started yet.
                contestOver:
                  value:
                    title: Submission rejected
                    status: 400
                    detail: contest is over.
        '403':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                unauthenticated:
                  value:
                    title: Forbidden
                    status: 403
                    detail: please authenticate before submit.
        '429':
          description: Submission rate limited
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - title
        - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    GameTickStatus:
      type: object
      additionalProperties: false
      required:
        - id
        - status
      properties:
        id:
          type: integer
          minimum: 0
        status:
          type: string
    GameSchedulerStatus:
      type: object
      additionalProperties: false
      required:
        - state
        - interval_seconds
      properties:
        state:
          type: string
        interval_seconds:
          type: integer
          minimum: 0
        last_run_at:
          type: string
          format: date-time
        next_run_at:
          type: string
          format: date-time
        last_tick_id:
          type: integer
          minimum: 0
        last_error:
          type: string
    GameMatchStatus:
      type: object
      additionalProperties: false
      required:
        - state
        - accepting_submissions
      properties:
        state:
          type: string
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        scheduled_start_at:
          type: string
          format: date-time
        scheduled_end_at:
          type: string
          format: date-time
        schedule_configured:
          type: boolean
        accepting_submissions:
          type: boolean
    GameStatus:
      type: object
      additionalProperties: false
      required:
        - total_ticks
      properties:
        match:
          $ref: '#/components/schemas/GameMatchStatus'
        current_tick:
          $ref: '#/components/schemas/GameTickStatus'
        scheduler:
          $ref: '#/components/schemas/GameSchedulerStatus'
        total_ticks:
          type: integer
          minimum: 0
...
    SubmitFlagsRequest:
      type: object
      additionalProperties: false
      required:
        - flags
      properties:
        flags:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
    SubmissionVerdict:
      type: object
      additionalProperties: false
      required:
        - flag
        - status
        - detail
      properties:
        flag:
          type: string
        status:
          type: string
          enum:
            - accepted
            - duplicate
            - invalid
        detail:
          type: string
          enum:
            - flag is correct.
            - flag already submitted.
            - flag is wrong or expired.
    SubmissionResult:
      type: object
      additionalProperties: false
      required:
        - results
        - accepted_count
        - rejected_count
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionVerdict'
        accepted_count:
          type: integer
          minimum: 0
        rejected_count:
          type: integer
          minimum: 0
...